requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
curl_cffi==0.7.4
distro==1.9.0
exceptiongroup==1.2.2
filetype==1.2.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
lxml==5.3.0
multidict==6.1.0
oauthlib==3.2.2
openai==0.28.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyotp==2.9.0
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.3.1
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
tomli==2.2.1
tqdm==4.67.1
tweepy==4.14.0
twikit==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
